// app/api/v2/leads/[id]/status/types.ts
export type RulesOnStatus =
   | 'Human Answered Status'
   | 'Ask Appointment Outcome'
   | 'Hidden Status'
   | 'Require Note'
   | 'Bucket Status'
   | 'Hide If Contacted'
   | 'Dollar Per Lead Status'
   | 'Scheduled Status'
   | 'Pipe Status'
   | 'Do Not Contact'
   | 'Required Number of Calls'
   | 'Trigger Webhook (beta)'
   | 'Remove From Pipe';